added shuffle to initalization
once the solution cards are added to each of their card type's decks, another shuffle is done of each.  Without the shuffle, when the arrays are printed for suggestions and accusations, the solution will always be the last on the list as it is the last added. The solution room will be at the end of the list of rooms.  Same with weapon and suspect