Add files via upload
This is the GUI semi-integrated with a websockets tutorial that I found on the internet (http://www.oracle.com/webfolder/technetwork/tutorials/obe/java/HomeWebsocket/WebsocketHome.html). What I did was putting my google on the project from that tutorial and played on how to implement the sockets. I managed to get something working on a chat. Getting the input from the user, converting it into a JSON type and then broadcast it back to all users. The methods that I used are commented. If you just want to see the GUI code, it is on the ClueGame\WebContent\index. The javascript for it is C:\Users\eliof\EEworkspace\ClueGame\WebContent\js\customjs. The GUI still has all the weird colors, but thats going to be removed.

The program can be run on the Glassfish server like the tutorial explained. You can try it by opening http://localhost:8080/WebsocketHome/ in a browser tab. Also on another tab open http://localhost:8080/WebsocketHome/indexPlayer2.html. You can now see how you're sending from