Trying to avoid conflicts
I spent a lot of time cleaning up the WebSocketServer.java and
GameSessionHandler.java files. I took out all the references to Devices
since it's not applicable to our project. Hopefully it will make the
code more readable. I also tried to add some additional functionality to
help with integrating sending messages from the game to users. I'm still
working on this aspect, but now that there's not a lot of clutter in the
code, I feel like it will go a lot faster.